+++ app/src/main/java/com/lucas/currencylist/ui/screens/currencies/CurrenciesViewModel.kt
@@ -3,8 +3,8 @@ package com.lucas.currencylist.ui.screens.currencies
 import android.app.Application
 import androidx.lifecycle.AndroidViewModel
 import com.lucas.core.data.local.ExchangeLocalDataSource
-import com.lucas.core.data.local.database.ExchangeValueDatabase
-import com.lucas.core.data.local.database.PlatformUpdatesDatabase
+import com.lucas.core.data.local.database.exchangeValues.ExchangeValueDatabase
+import com.lucas.core.data.local.database.platforms.PlatformUpdatesDatabase
 import com.lucas.core.data.remote.ExchangeRemoteDataSource
 import com.lucas.core.data.remote.apis.RetrofitBuilder
 import com.lucas.core.data.repositories.CurrencyRepository
