+++ core/src/main/java/com/lucas/core/data/local/database/exchangeValues/ExchangeValueDatabase.kt
@@ -1,34 +1,34 @@
-package com.lucas.core.data.local.database
+package com.lucas.core.data.local.database.exchangeValues
 
 import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
-import com.lucas.core.data.models.TradingPlatformUpdates
+import com.lucas.core.data.models.ExchangeValue
 import com.lucas.core.data.local.database.converters.DateTimeConverter
 
 
-@Database(entities = [TradingPlatformUpdates::class], version = 4, exportSchema = false)
+@Database(entities = [ExchangeValue::class], version = 1, exportSchema = false)
 @TypeConverters(DateTimeConverter::class)
-abstract class PlatformUpdatesDatabase : RoomDatabase() {
+abstract class ExchangeValueDatabase : RoomDatabase() {
 
-    abstract val dao: PlatformUpdatesDatabaseDAO
+    abstract val dao: ExchangeValuesDatabaseDAO
 
     companion object {
 
         @Volatile
-        private var INSTANCE: PlatformUpdatesDatabase? = null
+        private var INSTANCE: ExchangeValueDatabase? = null
 
-        fun getInstance(context: Context): PlatformUpdatesDatabase {
+        fun getInstance(context: Context): ExchangeValueDatabase {
             synchronized(this) {
                 var instance = INSTANCE
 
                 if (instance == null) {
                     instance = Room.databaseBuilder(
                         context.applicationContext,
-                        PlatformUpdatesDatabase::class.java,
-                        "platform_updates_database"
+                        ExchangeValueDatabase::class.java,
+                        "currencies_database"
                     )
                         .fallbackToDestructiveMigration()
                         .build()
