+++ core/src/main/java/com/lucas/core/data/local/database/platforms/PlatformUpdatesDatabase.kt
@@ -1,34 +1,34 @@
-package com.lucas.core.data.local.database
+package com.lucas.core.data.local.database.platforms
 
 import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
-import com.lucas.core.data.models.ExchangeValue
+import com.lucas.core.data.models.TradingPlatformUpdates
 import com.lucas.core.data.local.database.converters.DateTimeConverter
 
 
-@Database(entities = [ExchangeValue::class], version = 1, exportSchema = false)
+@Database(entities = [TradingPlatformUpdates::class], version = 4, exportSchema = false)
 @TypeConverters(DateTimeConverter::class)
-abstract class ExchangeValueDatabase : RoomDatabase() {
+abstract class PlatformUpdatesDatabase : RoomDatabase() {
 
-    abstract val dao: ExchangeValuesDatabaseDAO
+    abstract val dao: PlatformUpdatesDatabaseDAO
 
     companion object {
 
         @Volatile
-        private var INSTANCE: ExchangeValueDatabase? = null
+        private var INSTANCE: PlatformUpdatesDatabase? = null
 
-        fun getInstance(context: Context): ExchangeValueDatabase {
+        fun getInstance(context: Context): PlatformUpdatesDatabase {
             synchronized(this) {
                 var instance = INSTANCE
 
                 if (instance == null) {
                     instance = Room.databaseBuilder(
                         context.applicationContext,
-                        ExchangeValueDatabase::class.java,
-                        "currencies_database"
+                        PlatformUpdatesDatabase::class.java,
+                        "platform_updates_database"
                     )
                         .fallbackToDestructiveMigration()
                         .build()
