+++ core/src/main/java/com/lucas/core/domain/extensions/CurrencyExtensions.kt
@@ -1,6 +1,8 @@
 package com.lucas.core.domain.extensions
 
 import com.lucas.core.data.models.CurrencyType
+import com.lucas.core.data.models.ExchangePlatformType
+import com.lucas.core.data.models.ExchangeValue
 
 
 fun CurrencyType.getName(): String = when (this) {
@@ -12,4 +14,12 @@ fun CurrencyType.getName(): String = when (this) {
     CurrencyType.USDC -> "USDC"
     CurrencyType.BNB -> "BNB"
     else -> ""
+}
+
+fun List<ExchangeValue>.filterByPlatform(platform: ExchangePlatformType) = filter {
+    it.platform == platform
+}
+
+fun List<ExchangeValue>.anyPlatform(platform: ExchangePlatformType) = any {
+    it.platform == platform
 }
